refactor: address CR issues
